requirements_extra.txt
Sphinx>=1.8.5
numpydoc>=1.0.0
graphviz>=0.10.1
ipython>=5.5.0
pydata-sphinx-theme>=0.6.3
pytest>=3.6.4
